refactor: rename config changes history to config changes table
